+++ tests/Pest.php
@@ -3,3 +3,5 @@
 use CodebarAg\FlysystemCloudinary\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);
+
+uses()->group('integration')->in(__DIR__.'/Integration');